Computer vision using CV2
This is a simple computer vision program to detect moving objects with a well defined size. The size is defined to make sure the camera doesn't capture unnecessary objects like flies and start giving a warning sound, "It can be annoying". The will keep running and capturing objects until the letter "q" is pressed then it exits. 
My aim is to make it record and send suspicious movements to the owners mail id or phone number.

All contributions are welcomed.